refactoring header into more consistent cpp style
